extended detection of wine installation
